+++ Leviasan.Example.WebApplication/Program.cs
@@ -1,0 +1,36 @@
+using Leviasan.Sanlog.EntityFrameworkCore;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+internal sealed class Program
+{
+    private static void Main(string[] args)
+    {
+        var builder = WebApplication.CreateBuilder(args);
+        // Add services to the container.
+        builder.Services.AddControllers();
+        // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+        builder.Services.AddEndpointsApiExplorer();
+        builder.Services.AddSwaggerGen();
+        builder.Logging.AddSanlogLoggerEFCore(contextConfigure =>
+        {
+            contextConfigure.UseSqlServer("Server=(localdb)\\mssqllocaldb;Database=sanlogdb;Trusted_Connection=True");
+        });
+        builder.Services.AddHttpLogging(c => c.CombineLogs = true);
+
+        var app = builder.Build();
+        app.UseHttpLogging();
+        // Configure the HTTP request pipeline.
+        if (app.Environment.IsDevelopment())
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
+        }
+        app.UseHttpsRedirection();
+        app.UseAuthorization();
+        app.MapControllers();
+        app.Run();
+    }
+}