+++ Leviasan.Sanlog.EntityFrameworkCore/LoggingEntryConfiguration.cs
@@ -12,10 +12,11 @@ namespace Leviasan.Sanlog.EntityFrameworkCore
         public void Configure(EntityTypeBuilder<LoggingEntry> builder)
         {
             _ = builder.Property(x => x.Id).ValueGeneratedNever();
-            _ = builder.Property(x => x.Version).HasConversion<VersionValueConverter, VersionValueComparer>().HasMaxLength(43);
-            _ = builder.Property(x => x.Category).IsRequired(true);
-            _ = builder.Property(x => x.EventName).IsRequired(false);
-            _ = builder.Property(x => x.Message).IsRequired(false).HasMaxLength(int.MaxValue);
+            _ = builder.Property(x => x.Version).HasConversion<VersionValueConverter, VersionValueComparer>().IsRequired(false).IsUnicode(false).HasMaxLength(43);
+            _ = builder.Property(x => x.Category).IsRequired(true).IsUnicode(true);
+            _ = builder.Property(x => x.EventName).IsRequired(false).IsUnicode(true);
+            _ = builder.Property(x => x.Message).IsRequired(false).IsUnicode(true);
+            _ = builder.Property(x => x.Properties).HasConversion<StringDictionaryValueConverter, StringDictionaryValueComparer>().IsRequired(false).IsUnicode(true).HasMaxLength(int.MaxValue);
             _ = builder.HasIndex(x => x.DateTime);
         }
     }