+++ Leviasan.Sanlog.Abstractions/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2.0.0
 - MSBuild property `GenerateDocumentationFile` is enabled.
 - `FormattedLogValuesFormatter` implements `IFormatProvider` and `ICustomFormatter`.
-- `FormattedLogValuesFormatter` and `SanlogLoggerOptions` provide new API `RegisterSensitiveData(Type, string)` to register property whose value belongs to sensitive data. Using `string` redacts the property of the composite format string and `DictionaryEntry` redacts the dictionary entry value.
+- `FormattedLogValuesFormatter` and `SanlogLoggerOptions` provide new API `RegisterSensitiveData(Type, string)` to register property whose value belongs to sensitive data. Using `object` redacts the property of the composite format string and `DictionaryEntry` redacts the dictionary entry value.
 - Added new API to represent a message template `MessageTemplate`.
 
 ### 1.1.3